Initial extraction of sys components
