+++ AGENTS.md
@@ -19,6 +19,10 @@ Filename Conventions
 - Use explicit nouns for specialized renderers or formats (e.g., `ascii_image.rs`, `status_bar.rs`).
 - Do NOT include the word `component` in filenames (avoid `terminal_component.rs`).
 
+- Window-manager-specific components (in the `term-wm-sys-ui-components` crate) must use the `wm_` filename prefix and `Wm` type prefix.
+	- Example: `WmDebugLogComponent` -> `wm_debug_log.rs`, `WmMenuOverlay` -> `wm_menu_overlay.rs`.
+	- This applies to top-level files and files within category subdirectories (e.g., `sys/wm_help_overlay.rs`).
+	- Internal types (handles, writers, etc.) that are not components do not require the `Wm` prefix.
 
 Component Implementation Placement
 - The `impl Component for <Name>Component { ... }` block must appear immediately below the `struct <Name>Component { ... }` declaration in the same file.
