+++ crates/term-wm-sys-ui-components/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod sys;
+pub mod wm_dialog_overlay;
+pub mod wm_menu_overlay;
+
+pub use wm_dialog_overlay::WmDialogOverlayComponent;
+pub use wm_menu_overlay::WmMenuOverlay;
+pub use sys::*;
