+++ crates/term-wm-sys-ui-components/src/sys/mod.rs
@@ -1,0 +1,3 @@
+pub mod wm_debug_log;
+pub mod wm_help_overlay;
+pub mod wm_keybinding_overlay;
