+++ crates/term-wm-sys-ui-components/src/sys/wm_debug_log.rs
@@ -6,11 +6,11 @@ use crossterm::event::Event;
 use ratatui::layout::Rect;
 use ratatui::text::{Line, Text};
 
-use crate::{ScrollViewComponent, TextRendererComponent};
 use term_wm_core::components::{Component, ComponentContext};
 use term_wm_core::debug_event_flags;
 use term_wm_core::ui::UiFrame;
 use term_wm_core::window::{SystemWindowView, WindowSurface};
+use term_wm_ui_components::{ScrollViewComponent, TextRendererComponent};
 
 const DEFAULT_MAX_LINES: usize = 2000;
 static GLOBAL_LOG: OnceLock<DebugLogHandle> = OnceLock::new();
@@ -59,7 +59,6 @@ pub fn install_panic_hook() {
             }
             handle.push("============".to_string());
         }
-        // Mark that a panic occurred so the UI can react in the next frame.
         debug_event_flags::trigger_panic_pending();
         prev(info);
     }));
@@ -159,7 +158,7 @@ impl Write for DebugLogWriter {
 }
 
 #[derive(Debug)]
-pub struct DebugLogComponent {
+pub struct WmDebugLogComponent {
     handle: DebugLogHandle,
     scroll_view: ScrollViewComponent<TextRendererComponent>,
     follow_tail: bool,
@@ -167,25 +166,11 @@ pub struct DebugLogComponent {
     last_view: usize,
 }
 
-impl Component for DebugLogComponent {
+impl Component for WmDebugLogComponent {
     fn render(&mut self, frame: &mut UiFrame<'_>, area: Rect, ctx: &ComponentContext) {
         if area.width == 0 || area.height == 0 {
             return;
         }
-        let buffer = frame.buffer_mut();
-        let bounds = area.intersection(buffer.area);
-        if bounds.width == 0 || bounds.height == 0 {
-            return;
-        }
-        for y in bounds.y..bounds.y.saturating_add(bounds.height) {
-            for x in bounds.x..bounds.x.saturating_add(bounds.width) {
-                if let Some(cell) = buffer.cell_mut((x, y)) {
-                    cell.reset();
-                    cell.set_symbol(" ");
-                }
-            }
-        }
-        // Build text from the handle buffer
         let lines = if let Ok(buffer) = self.handle.inner.lock() {
             buffer.lines.iter().cloned().collect::<Vec<_>>()
         } else {
@@ -196,28 +181,20 @@ impl Component for DebugLogComponent {
         self.last_total = total;
         self.last_view = view;
 
-        // Prepare Text for the renderer
         let text = Text::from(lines.into_iter().map(Line::from).collect::<Vec<_>>());
         self.scroll_view.content.set_text(text);
         self.scroll_view.content.set_wrap(false);
-        // Follow tail behavior: set renderer offset to bottom when enabled
         if self.follow_tail {
-            // renderer will position itself to the bottom during render via its internal scroll
             self.scroll_view
                 .viewport_handle()
                 .scroll_vertical_to(usize::MAX);
         }
-        // We need to render first to update scrollbar/offsets if we want is_at_bottom to reflect *after* render state?
-        // But original code did `follow_tail = is_at_bottom()` BEFORE render (after explicit set offset).
-        // If we just set offset to MAX, then is_at_bottom should be true.
-        // But internal `is_at_bottom` reads `renderer_offset`. `ViewportHandle` updates immediately.
         self.follow_tail = self.is_at_bottom();
 
         self.scroll_view.render(frame, area, ctx);
     }
 
     fn handle_event(&mut self, event: &Event, ctx: &ComponentContext) -> bool {
-        // Delegate to the renderer which handles scroll/key/mouse
         let handled = self.scroll_view.handle_event(event, ctx);
         if handled {
             self.follow_tail = self.is_at_bottom();
@@ -226,14 +203,13 @@ impl Component for DebugLogComponent {
     }
 }
 
-impl DebugLogComponent {
+impl WmDebugLogComponent {
     pub fn new(max_lines: usize) -> (Self, DebugLogHandle) {
         let handle = DebugLogHandle {
             inner: Arc::new(Mutex::new(DebugLogBuffer::new(max_lines))),
         };
         let mut renderer = TextRendererComponent::new();
         renderer.set_wrap(false);
-        // set_keyboard_enabled removed
         (
             Self {
                 handle: handle.clone(),
@@ -267,7 +243,7 @@ impl DebugLogComponent {
     }
 }
 
-impl SystemWindowView for DebugLogComponent {
+impl SystemWindowView for WmDebugLogComponent {
     fn render(
         &mut self,
         frame: &mut term_wm_core::ui::UiFrame<'_>,
@@ -275,7 +251,7 @@ impl SystemWindowView for DebugLogComponent {
         focused: bool,
     ) {
         let ctx = ComponentContext::new(focused);
-        <DebugLogComponent as Component>::render(self, frame, surface.inner, &ctx);
+        <WmDebugLogComponent as Component>::render(self, frame, surface.inner, &ctx);
     }
 
     fn handle_event(&mut self, event: &Event) -> bool {
@@ -283,7 +259,7 @@ impl SystemWindowView for DebugLogComponent {
     }
 
     fn set_selection_enabled(&mut self, enabled: bool) {
-        DebugLogComponent::set_selection_enabled(self, enabled);
+        WmDebugLogComponent::set_selection_enabled(self, enabled);
     }
 }
 
@@ -297,12 +273,11 @@ mod tests {
 
     #[test]
     fn debug_log_handle_and_buffer_limits() {
-        let (_comp, handle) = DebugLogComponent::new(3);
+        let (_comp, handle) = WmDebugLogComponent::new(3);
         handle.push("one");
         handle.push("two");
         handle.push("three");
         handle.push("four");
-        // internal buffer should be capped at 3
         if let Ok(buf) = handle.inner.lock() {
             assert_eq!(buf.lines.len(), 3);
             assert_eq!(buf.lines.front().unwrap().as_str(), "two");
@@ -313,10 +288,9 @@ mod tests {
 
     #[test]
     fn debug_log_writer_flushes_lines() {
-        let (_comp, handle) = DebugLogComponent::new(10);
+        let (_comp, handle) = WmDebugLogComponent::new(10);
         let mut writer = handle.writer();
         let _ = writer.write(b"first line\nsecond line\npartial");
-        // flush should push pending partial when forced
         writer.flush().unwrap();
         if let Ok(buf) = handle.inner.lock() {
             assert!(buf.lines.iter().any(|s| s == "first line"));
@@ -327,7 +301,7 @@ mod tests {
 
     #[test]
     fn debug_log_component_handle_event_scrolls() {
-        let (mut comp, handle) = DebugLogComponent::new(10);
+        let (mut comp, handle) = WmDebugLogComponent::new(10);
         for i in 0..20 {
             handle.push(format!("line{i}"));
         }
