+++ crates/term-wm-ui-components/src/lib.rs
@@ -3,12 +3,10 @@ pub mod confirm_overlay;
 pub mod dialog_overlay;
 pub mod list;
 pub mod menu;
-pub mod menu_overlay;
 pub mod panel;
 pub mod markdown_viewer;
 pub mod scroll_view;
 pub mod svg_image;
-pub mod sys;
 pub mod terminal;
 pub mod text_renderer;
 pub mod toggle_list;
@@ -18,14 +16,12 @@ pub use confirm_overlay::ConfirmOverlayComponent;
 pub use dialog_overlay::DialogOverlayComponent;
 pub use list::ListComponent;
 pub use menu::MenuComponent;
-pub use menu_overlay::WmMenuOverlay;
 pub use panel::PanelComponent;
 pub use markdown_viewer::MarkdownViewerComponent;
 pub use scroll_view::{
     ScrollViewComponent, ScrollbarAxis, ScrollbarDrag, render_scrollbar, render_scrollbar_oriented,
 };
 pub use svg_image::SvgImageComponent;
-pub use sys::*;
 pub use terminal::{TerminalComponent, default_shell, default_shell_command};
 pub use text_renderer::TextRendererComponent;
 pub use toggle_list::{ToggleItem, ToggleListComponent};
