--- crates/term-wm-ui-components/src/sys/mod.rs
@@ -1,3 +1,0 @@
-pub mod debug_log;
-pub mod help_overlay;
-pub mod keybinding_overlay;
