+++ examples/dual_image.rs
@@ -60,7 +60,7 @@ impl App {
                 "example", "0.0.0", hostname,
             ));
         let menu_overlay: Box<dyn term_wm_core::components::MenuOverlay<term_wm_core::window::WmMenuAction>> =
-            Box::new(term_wm_ui_components::WmMenuOverlay::new());
+            Box::new(term_wm_sys_ui_components::WmMenuOverlay::new());
         let mut windows = WindowManager::with_config(
             PaneId::Left,
             WmConfig::standalone(),
