+++ src/main.rs
@@ -22,8 +22,8 @@ use term_wm::ui::UiFrame;
 use term_wm::window::{OverlayId, SystemWindowId, WindowDrawContext, WindowManager};
 use term_wm::wm_config::WmConfig;
 use term_wm::{ScrollViewComponent, TerminalComponent, default_shell_command};
-use term_wm_ui_components::sys::debug_log::{
-    DebugLogComponent, install_panic_hook, set_global_debug_log,
+use term_wm_sys_ui_components::sys::wm_debug_log::{
+    WmDebugLogComponent, install_panic_hook, set_global_debug_log,
 };
 
 type PaneId = usize;
@@ -105,7 +105,7 @@ impl App {
                 hostname,
             ));
         let config = WmConfig::standalone();
-        let mut raw_menu = term_wm_ui_components::WmMenuOverlay::new();
+        let mut raw_menu = term_wm_sys_ui_components::WmMenuOverlay::new();
         raw_menu.set_timeout(config.menu_outline_timeout);
         let menu_overlay: Box<dyn MenuOverlay<term_wm_core::window::WmMenuAction>> =
             Box::new(raw_menu);
@@ -122,7 +122,7 @@ impl App {
 
         // Initialize debug log system window
         {
-            let (mut component, handle) = DebugLogComponent::new_default();
+            let (mut component, handle) = WmDebugLogComponent::new_default();
             component.set_selection_enabled(app.windows.clipboard_enabled());
             set_global_debug_log(handle);
             app.windows
@@ -200,18 +200,18 @@ impl WindowManagerHost<PaneId> for App {
     }
 
     fn open_help_overlay(&mut self) {
-        use term_wm_ui_components::sys::help_overlay::HelpOverlayComponent;
+        use term_wm_sys_ui_components::sys::wm_help_overlay::WmHelpOverlayComponent;
         let kb = self.windows.keybindings().clone();
-        let mut h = HelpOverlayComponent::new(kb);
+        let mut h = WmHelpOverlayComponent::new(kb);
         h.show();
         h.set_selection_enabled(self.windows.clipboard_enabled());
         self.windows.open_overlay(OverlayId::Help, Box::new(h));
     }
 
     fn open_keybindings_overlay(&mut self) {
-        use term_wm_ui_components::sys::keybinding_overlay::KeybindingOverlayComponent;
+        use term_wm_sys_ui_components::sys::wm_keybinding_overlay::WmKeybindingOverlayComponent;
         let kb = self.windows.keybindings().clone();
-        let mut o = KeybindingOverlayComponent::new(kb);
+        let mut o = WmKeybindingOverlayComponent::new(kb);
         o.show();
         self.windows
             .open_overlay(OverlayId::Keybindings, Box::new(o));
