+++ tests/integration_new_window_focus.rs
@@ -10,7 +10,7 @@ fn new_window_is_focused() {
             &ctx.app_name, &ctx.app_version, None,
         ));
     let menu: Box<dyn term_wm_core::components::MenuOverlay<term_wm_core::window::WmMenuAction>> =
-        Box::new(term_wm_ui_components::WmMenuOverlay::new());
+        Box::new(term_wm_sys_ui_components::WmMenuOverlay::new());
     let mut wm = WindowManager::<usize>::with_config(
         0,
         term_wm::wm_config::WmConfig::standalone(),
